Do not allow matching on empty string
also addresses other copilot review feedback.
