+++ cedar-policy-core/src/parser/grammar.lalrpop
@@ -35,7 +35,7 @@ extern {
 
 match {
     // Whitespace and comments
-    r"\s*" => { }, // The default whitespace skipping is disabled an `ignore pattern` is specified
+    r"\s+" => { }, // The default whitespace skipping is disabled an `ignore pattern` is specified
     r"//[^\n\r]*[\n\r]*" => { }, // Skip `// comments`
 
     // Special Identifiers (begin expressions)
