+++ cedar-policy-core/src/validator/cedar_schema/grammar.lalrpop
@@ -60,7 +60,7 @@ extern {
 
 match {
     // Whitespace and comments
-    r"\s*" => { }, // The default whitespace skipping is disabled an `ignore pattern` is specified
+    r"\s+" => { }, // The default whitespace skipping is disabled an `ignore pattern` is specified
     r"//[^\n\r]*[\n\r]*" => { }, // Skip `// comments`
 
     // keywords
